huge change
When the codebase is too big, it triggers relocation error for
functions to find bb_hook when using position-indepdenent code
if the whole code is in one object file.
Therefore, now the library allows us to extract the helper function
out of the codebase and link the helper through the end of the
linker stage.